refactor: switch to combinators in `data` module & code move + update imports
