+++ src/strategy/mod.rs
@@ -2,7 +2,7 @@ pub mod error;
 pub mod signal;
 pub mod strategy;
 
-use crate::data::market::MarketEvent;
+use crate::data::MarketEvent;
 use crate::strategy::signal::SignalEvent;
 
 /// May generate an advisory [`SignalEvent`] as a result of analysing an input [`MarketEvent`].
